[dev] use CMC when token is absent in strategy_pool_asset_last_prices_in_base_token

+++ src/service/user/pg_func.rs
@@ -1973,7 +1973,7 @@ END
                 Field::new("short_name", Type::String),
                 Field::new("description", Type::String),
                 Field::new("is_stablecoin", Type::Boolean),
-                Field::new("decimals", Type::Numeric),
+                Field::new("decimals", Type::Int),
             ],
             r#"
 BEGIN
